+++ tests/0300_views/0313_category.py
@@ -11,10 +11,36 @@ from atoum.utils.tests import html_pyquery
 from tests.initial import initial_catalog  # noqa: F401
 
 
+def test_anonymous(client, db, settings):
+    """
+    Anonymous are not allowed and is redirect to login.
+    """
+    url = reverse("atoum:category-index")
+    response = client.get(url, follow=True)
+    assert response.redirect_chain == [
+        ("{}?next={}".format(settings.LOGIN_URL, url), 302),
+    ]
+    assert response.status_code == 200
+
+    category = CategoryFactory()
+    url = reverse("atoum:category-detail", kwargs={
+        "consumable_slug": category.assortment.consumable.slug,
+        "assortment_slug": category.assortment.slug,
+        "category_slug": category.slug,
+    })
+    response = client.get(url, follow=True)
+    assert response.redirect_chain == [
+        ("{}?next={}".format(settings.LOGIN_URL, url), 302),
+    ]
+    assert response.status_code == 200
+
+
 def test_index_empty(client, db):
     """
     Category index should just respond with an empty list.
     """
+    client.force_login(UserFactory())
+
     url = reverse("atoum:category-index")
     response = client.get(url, follow=True)
     assert response.redirect_chain == []
@@ -29,11 +55,13 @@ def test_index_filled(client, db, initial_catalog,  # noqa: F811
     """
     Category index should list available assortments with pagination.
     """
+    client.force_login(UserFactory())
+
     url = reverse("atoum:category-index")
 
     # Only a count queryset from pagination and the other one to list objects with
     # their relation preloaded
-    with django_assert_num_queries(2):
+    with django_assert_num_queries(4):
         response = client.get(url, follow=True)
 
     assert response.redirect_chain == []
@@ -60,6 +88,8 @@ def test_detail_filled(client, db, initial_catalog,  # noqa: F811
     """
     Category detail should list its related products.
     """
+    client.force_login(UserFactory())
+
     beeffoods = initial_catalog.categories["beeffoods"]
     url = reverse(
         "atoum:category-detail",
@@ -72,7 +102,7 @@ def test_detail_filled(client, db, initial_catalog,  # noqa: F811
 
     # A queryset for the main object, another one to list its related objects and
     # another one for pagination
-    with django_assert_num_queries(3):
+    with django_assert_num_queries(5):
         response = client.get(url, follow=True)
 
     assert response.redirect_chain == []