+++ tests/0300_views/0315_product.py
@@ -12,10 +12,37 @@ from atoum.utils.tests import html_pyquery
 from tests.initial import initial_catalog  # noqa: F401
 
 
+def test_anonymous(client, db, settings):
+    """
+    Anonymous are not allowed and is redirect to login.
+    """
+    url = reverse("atoum:product-index")
+    response = client.get(url, follow=True)
+    assert response.redirect_chain == [
+        ("{}?next={}".format(settings.LOGIN_URL, url), 302),
+    ]
+    assert response.status_code == 200
+
+    product = ProductFactory()
+    url = reverse("atoum:product-detail", kwargs={
+        "consumable_slug": product.category.assortment.consumable.slug,
+        "assortment_slug": product.category.assortment.slug,
+        "category_slug": product.category.slug,
+        "product_slug": product.slug,
+    })
+    response = client.get(url, follow=True)
+    assert response.redirect_chain == [
+        ("{}?next={}".format(settings.LOGIN_URL, url), 302),
+    ]
+    assert response.status_code == 200
+
+
 def test_index_empty(client, db):
     """
     Product index should just respond with an empty list.
     """
+    client.force_login(UserFactory())
+
     url = reverse("atoum:product-index")
     response = client.get(url, follow=True)
     assert response.redirect_chain == []
@@ -30,11 +57,13 @@ def test_index_filled(client, db, initial_catalog,  # noqa: F811
     """
     Product index should list available assortments with pagination.
     """
+    client.force_login(UserFactory())
+
     url = reverse("atoum:product-index")
 
     # Only a count queryset from pagination and the other one to list objects with
     # their relation preloaded
-    with django_assert_num_queries(2):
+    with django_assert_num_queries(4):
         response = client.get(url, follow=True)
 
     assert response.redirect_chain == []